perf: improve performance of `String.ValidPos` (#11142)
This PR aims to bring the performance of `String.ValidPos` closer to
that of `String.Pos.Raw` by adding/correcting `extern` annotations as
needed.

This is in response to a regression observed after #11127. The changes
to the `String` `Parsec` module lead to different compiler behavior for
functions like `strCore` and `natCore`. The new IR *looks* better than
the old IR, but the
[numbers](https://radar.lean-lang.org/repos/lean4/commits/1e438647bae47ea623cc12897010fd68f37d5884?reference=d1e19f2aa0eb4c9b9c8684d80389106c3d21db8b)
are a bit mixed.