feat: add `BitVec.getLsbD_false_of_clz`
This PR adds `BitVec.getLsbD_false_of_clz`, which states that the boolean value
of any bit in the 'leading zero bits' of a BitVec is false.
`BitVec.getLsbD_false_of_clzAuxRec` does already exist, but `BitVec.clzAuxRec`
is really just an implementation detail of `BitVec.clz`, and it is more useful
to have lemmas which directly state properties of `BitVec.clz`, rather than
the auxiliary function which it uses internally.