doc: write a guideline for tactic docstrings
At the Mathlib Initiative we want to ensure that tactics have good documentation. Since this will involve adding documentation to tactics built into core Lean, I discussed with David that we should have a shared set of documentation guidelines that allow me to do my work both on the Lean and on the Mathlib repositories.

I have already shown an earlier version of this guideline to David who made some helpful suggestions but would be away. So to make sure the discussion doesn't get lost, I've made a PR with the version I ended up with after the first round of comments.