feat: use approximate universe inverse when `sorry` in inductive type
This PR modifies the universe level constraint algorithms for the inductive type elaborator to use an approximation for constructor fields that contain sorry. For example, the following used to fail with an 'accidental higher universe' error due to the fact that the `sorry`'s universe level constraint is of the form `u ≤ ?r + 1`, which is often a mistake.
```lean4
inductive Sorry1 where
  | x (a : Array Sorry1)
  | y (b : sorry)
  ```
The following example also used to fail, since the constraint was `?u ≤ 1`, which does not have a unique solution. It now takes the least solution if there is a `sorry`.
  ```lean4
  inductive Sorry2 : Type where
  | y (b : sorry)
  ```